// src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  OWNER: 'OWNER',
  MANAGER: 'MANAGER',
  STAFF: 'STAFF'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const ProductCategory = {
  WHISKY: 'WHISKY',
  RUM: 'RUM',
  VODKA: 'VODKA',
  GIN: 'GIN',
  BRANDY: 'BRANDY',
  WINE: 'WINE',
  BEER: 'BEER'
} as const

export type ProductCategory = (typeof ProductCategory)[keyof typeof ProductCategory]


export const SizeUnit = {
  ML_750: 'ML_750',
  ML_375: 'ML_375',
  ML_180: 'ML_180',
  CAN_500: 'CAN_500',
  BOTTLE_650: 'BOTTLE_650'
} as const

export type SizeUnit = (typeof SizeUnit)[keyof typeof SizeUnit]


export const PaymentMode = {
  CASH: 'CASH',
  UPI: 'UPI',
  CREDIT: 'CREDIT'
} as const

export type PaymentMode = (typeof PaymentMode)[keyof typeof PaymentMode]